+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0]
 ### Added
+- Empty state enhancements ([#22](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/22))
+    - No match found
+    - Spinner
 - Expandable cards ([#10](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/10))
 - Spinner component ([#51](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/51))  
 - Text component ([#53](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/53))
@@ -13,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Vertical divider ([#19](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/19))
 
 ### Changed
+- Documentation for empty state component ([#22](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/22))
 - Documentation for pagination component ([#69](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/69))
 - Documentation for switch component ([#70](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/70))
 - Documentation for tab component ([#71](https://github.com/patternfly-kotlin/patternfly-fritz2/issues/71))